fixed a race condition with stopping the instance;
Starting the instance was not a problem, becuase that was less
sensitive to timing (though I did miss more boot-time console output
than I would've liked).

However, stopping the instance wasn't working, because it would
already be shut down in most of the cases, before the setup fo the
console could complete (or often before it could even start).

So I set about replicating the concurrency logic I implemented with
the shell one-liners in "crib-sheet.txt".  This requried using the
threading library, which was far less painful than I was expecting.

I could also make the fudge-factor-delay the same in both cases.

Now, seeing as both the stop and start scripts are nearly identical,
but for two lines, I think a prudent enhancement would be to combine
them into one, with a parameter to specifiy which behavior to take.

Another enhancement would be (after lopping-off the first arg) to pass
sys.argv directly to any arg that takes "[ instanceId ]" as input,
trivially expanding this to operate on multiple instances.