a test of creating an instance, with password creation via UserData;
It wasn't really necessary to write-up this test case for the console,
but I did it because I wanted to try using crypt(3) to generate a hash
for an account creation script that I pass in via the UserData
mechanism.

Now, since I've made an instance creation script, I should also
provide an instance destroy script as well.